Refactored plugin name documenter to documentor
