+++ ac.soton.eventb.documenter/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>ac.soton.eventb.documenter</name>
+	<name>ac.soton.eventb.documentor</name>
 	<comment></comment>
 	<projects>
 	</projects>
