+++ ac.soton.eventb.documenter/src/ac/soton/eventb/documentor/DiagramExporter.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     University of Southampton - initial API and implementation
  *******************************************************************************/
-package ac.soton.eventb.documenter;
+package ac.soton.eventb.documentor;
 
 import java.util.Map;
 import org.eclipse.core.resources.IFile;
