+++ ac.soton.eventb.documenter/src/ac/soton/eventb/documentor/StyleFile.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     University of Southampton - initial API and implementation
  *******************************************************************************/
-package ac.soton.eventb.documenter;
+package ac.soton.eventb.documentor;
 
 import java.io.InputStream;
